+++ test/inquery/query/chainable_test.rb
@@ -74,6 +74,72 @@ module Inquery
           query_class.run(Group.all)
         end
       end
+
+      def test_chainable_with_default_relation
+        query_class = Class.new(Inquery::Query::Chainable) do
+          relation class: 'Group'
+
+          def call
+            relation.where(color: 'red')
+          end
+        end
+
+        # Should use default relation when no relation is passed
+        result = query_class.run
+        assert_equal Group.find([1]), result.to_a
+      end
+
+      def test_chainable_with_params_only
+        result = Queries::Group::FilterWithColor.run(color: 'red')
+        assert_equal Group.find([1]), result.to_a
+      end
+
+      def test_chainable_with_empty_relation
+        query_class = Class.new(Inquery::Query::Chainable) do
+          relation class: 'Group'
+
+          def call
+            relation.where(color: 'blue')
+          end
+        end
+
+        result = query_class.run(Group.where('1=0'))
+        assert_empty result.to_a
+      end
+
+      def test_chainable_with_different_call_signatures
+        query_class = Class.new(Inquery::Query::Chainable) do
+          relation class: 'Group'
+          schema3 do
+            str? :color
+          end
+
+          def call
+            if osparams.color
+              relation.where(color: osparams.color)
+            else
+              relation
+            end
+          end
+        end
+
+        # Test with relation and params
+        result1 = query_class.run(Group.all, color: 'red')
+        assert_equal Group.find([1]), result1.to_a
+
+        # Test with params only
+        result2 = query_class.run(color: 'green')
+        assert_equal Group.find([2, 3]), result2.to_a
+
+        # Test with relation only (should return all)
+        result3 = query_class.run(Group.where('id > 1'))
+        assert_equal Group.find([2, 3]), result3.to_a
+      end
+
+      def test_relation_returns_active_record_relation
+        result = Queries::Group::FetchRed.run
+        assert_kind_of ActiveRecord::Relation, result
+      end
     end
   end
 end